+++ example/formatter/stop.js
@@ -4,7 +4,7 @@ class Stop {
 
         const parts = [];
 
-        parts.push(`[${event.server}]`);
+        parts.push(`[${event.serverInfo.shortcut}]`);
         parts.push(`${event.username}`);
         parts.push(`im Dungeon ${event.game_id}`);