+++ src/translator.js
@@ -2,11 +2,11 @@ const Events = require('./events.js');
 
 class Translator {
 
-    constructor(/* String */ shortcut, /* EventEmitter */ emitter, /* Store */ store) {
+    constructor(/* ServerInfo */ serverInfo, /* EventEmitter */ emitter, /* Store */ store) {
 
-        this.shortcut = shortcut;
-        this.store    = store;
-        this.emitter  = emitter;
+        this.serverInfo = serverInfo;
+        this.store      = store;
+        this.emitter    = emitter;
 
         this.ignoreMessages = [
             'Ping',
@@ -25,14 +25,14 @@ class Translator {
 
             switch (message.constructor.name) {
             case 'LobbyRemove':
-                this.emit('stop', new Events.Stop(this.shortcut, this.store.getById(message.id), message));
+                this.emit('stop', new Events.Stop(this.serverInfo, this.store.getById(message.id), message));
                 this.store.removeById(message.id);
                 break;
             case 'LobbyEntry':
                 if (this.store.hasEntry(message.game_id, message.username)) {
-                    this.emit('update', new Events.Update(this.shortcut, this.store.getEntry(message.game_id, message.username), message));
+                    this.emit('update', new Events.Update(this.serverInfo, this.store.getEntry(message.game_id, message.username), message));
                 } else {
-                    this.emit('start', new Events.Start(this.shortcut, message));
+                    this.emit('start', new Events.Start(this.serverInfo, message));
                 }
 
                 this.store.push(message);