+++ Changelog.md
@@ -1,3 +1,11 @@
+# 1.1.0
+
+Example: Start event formatter creates now watch url.
+ServerInfo: Lobby expects now a server url. WebSocket-Url and Watch-Url can then computed.
+Sinks: You can now add multiple sinks.
+Example: spectator info removed. It was to spammy.
+Example: Superfluous space in stop formatter removed.
+
 # 1.0.2
 
 Fixed typo in README.md